実行タイミングの調整
